Backend/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
pyjwt
python-dotenv
asyncpg
python-jose[cryptography]
pydantic[email]
bcrypt==4.1.2

